+++ spec/alpha_card/objects/capture_spec.rb
@@ -14,6 +14,10 @@ describe AlphaCard::Capture do
   context 'with valid attributes' do
     let(:capture) { AlphaCard::Capture.new(transaction_id: 'Some ID', amount: '10.05', order_id: '1', shipping_carrier: '2') }
 
+    let(:order) { AlphaCard::Order.new(id: '1', description: 'Test') }
+    let(:card_exp) { "#{'%02d' % Time.now.month}/#{Time.now.year.next}" }
+    let(:sale) { AlphaCard::Sale.new(card_expiration_date: card_exp, card_number: '4111111111111111', amount: '5.00') }
+
     it 'has valid request params' do
       expected_params = {
         transactionid: 'Some ID',
@@ -24,6 +28,17 @@ describe AlphaCard::Capture do
       }
 
       expect(capture.attributes_for_request).to eq(expected_params)
+    end
+
+    it 'processed successfully' do
+      allow_any_instance_of(AlphaCard::Sale).to receive(:type).and_return('auth')
+      success, response = sale.create(order, account)
+      expect(success).to be_truthy
+      expect(response.transaction_id).not_to be_nil
+
+      success, response = AlphaCard::Capture.new(transaction_id: response.transaction_id, amount: '2.00').create(account)
+      expect(success).to be_truthy
+      expect(response.text).to eq('SUCCESS')
     end
   end