+++ spec/spec_helper.rb
@@ -1,7 +1,12 @@
 require 'rack/test'
 
-require 'coveralls'
-Coveralls.wear!
+if ENV['CI'] || ENV['TRAVIS'] || ENV['COVERALLS'] || ENV['JENKINS_URL']
+  require 'coveralls'
+  Coveralls.wear!
+else
+  require 'simplecov'
+  SimpleCov.start
+end
 
 require 'bundler/setup'
 Bundler.setup